Add styles support
