+++ src/DragAndDrop/Draggable/Draggable.js
@@ -1,182 +1,25 @@
 import React from 'react';
 import WixComponent from '../../BaseComponents/WixComponent';
-import DragLayer from './DragLayer';
 import PropTypes from 'prop-types';
-import {DragSource as _DragSource, DropTarget as _DropTarget} from 'react-dnd';
-import {dragCoordinates} from './DragUtils';
-import {getEmptyImage} from 'react-dnd-html5-backend';
-import noop from 'lodash/noop';
+import DraggableSource from './components/DraggableSource';
+import DraggableTarget from './components/DraggableTarget';
 
-const ItemTypes = {
-  DRAGGABLE: 'Draggable'
-};
-
-const source = {
-  beginDrag: ({id, index, containerId, groupName, item, onMoveOut}) => {
-    return {
-      id,
-      index,
-      containerId,
-      groupName,
-      originalItem: item,
-      onMoveOut
-    };
-  },
-  endDrag: ({item, index, containerId, onDrop}, monitor) => {
-    if (monitor.didDrop()) {
-      onDrop({
-        payload: item,
-        removedIndex: index,
-        addedIndex: monitor.getDropResult().index,
-        addedToContainerId: monitor.getDropResult().containerId,
-        removedFromContainerId: containerId
-      });
-    }
-  },
-  isDragging: ({id, containerId, groupName}, monitor) => {
-    const item = monitor.getItem();
-    const isSameGroup = groupName && item.groupName && groupName === item.groupName;
-    const isSameContainer = containerId === item.containerId;
-    return (isSameGroup || isSameContainer) && item.id === id;
-  }
-};
-
-@_DragSource(ItemTypes.DRAGGABLE, source, (connect, monitor) => ({
-  connectDragSource: connect.dragSource(),
-  connectDragPreview: connect.dragPreview(),
-  isDragging: monitor.isDragging()
-}))
-export class DraggableSource extends React.Component {
-  componentDidMount() {
-    if (this.props.connectDragPreview) {
-      this.props.connectDragPreview(getEmptyImage());
-    }
-  }
-
-  _renderDraggableItem() {
-    const {isDragging, connectDragSource, withHandle, renderItem, id, item} = this.props;
-    if (withHandle) {
-      return renderItem({
-        id,
-        item,
-        isPlaceholder: isDragging,
-        connectHandle: handle => connectDragSource(handle)
-      });
-    }
-
-    return connectDragSource(
-      renderItem({
-        id,
-        item,
-        isPlaceholder: isDragging,
-        connectHandle: noop
-      })
-    );
-  }
-
-  _renderPreviewItem() {
-    const {renderItem, id, item} = this.props;
-    return (
-      <DragLayer
-        renderPreview={() => renderItem({
-          id,
-          item,
-          isPreview: true,
-          connectHandle: noop
-        })}
-        id={id}
-        draggedType={ItemTypes.DRAGGABLE}
-        />
-    );
-  }
-
-  render() {
-    const {connectDragSource} = this.props;
-    return connectDragSource ? (
-      <div>
-        {this._renderDraggableItem()}
-        {this._renderPreviewItem()}
-      </div>
-    ) : null;
-  }
-}
-
-DraggableSource.propTypes = {
-  isDragging: PropTypes.bool, // from react-dnd
-  connectDragSource: PropTypes.func, // from react-dnd
-  connectDragPreview: PropTypes.func, // from react-dnd
-
-  groupName: PropTypes.string,
-  containerId: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  renderItem: PropTypes.func,
-  index: PropTypes.number,
-  id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  item: PropTypes.object,
-  withHandle: PropTypes.bool,
-  onDrop: PropTypes.func,
-  onHover: PropTypes.func,
-  onMoveOut: PropTypes.func
-};
-
-const target = {
-  drop(props) {
-    return {
-      containerId: props.containerId,
-      index: props.index
-    };
-  },
-  hover(props, monitor, component) {
-    const monitorItem = monitor.getItem();
-    const dragIndex = monitorItem.index;
-    const hoverIndex = props.index;
-
-    const isSameGroup = props.groupName && monitorItem.groupName && props.groupName === monitorItem.groupName;
-    const isSameContainer = props.containerId === monitor.getItem().containerId;
-
-    if (!isSameContainer && !isSameGroup) {
-      return;
-    }
-
-    if (!component || (hoverIndex === dragIndex && isSameContainer)) {
-      return;
-    }
-
-    const {hoverClientY, hoverMiddleY} = dragCoordinates({monitor, component});
-    if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-      console.log(dragIndex, hoverIndex, hoverClientY, hoverMiddleY);
-      return;
-    }
-    if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-      console.log(dragIndex, hoverIndex, hoverClientY, hoverMiddleY);
-      return;
-    }
-
-    if (isSameGroup && !isSameContainer) {
-      monitorItem.onMoveOut(monitorItem.id);
-    }
-
-    props.onHover(dragIndex, hoverIndex, {
-      id: monitorItem.id,
-      item: monitorItem.originalItem,
-      type: isSameGroup && !isSameContainer ? 'group' : 'container'
-    });
-    monitor.getItem().index = hoverIndex;
-  }
-};
-
-@_DropTarget(ItemTypes.DRAGGABLE, target, connect => ({
-  connectDropTarget: connect.dropTarget()
-}))
 export class Draggable extends WixComponent {
   render() {
-    const {connectDropTarget, ...props} = this.props;
-    return connectDropTarget ?
-      connectDropTarget(
-        <div>
-          <DraggableSource {...props}/>
-        </div>
-      ) :
-      null;
+    return (
+      <DraggableTarget
+        {...this.props}
+        className={this.props.classes.dragTarget}
+        >
+        <DraggableSource
+          {...this.props}
+          classes={{
+            root: this.props.classes.dragSource,
+            dragLayer: this.props.classes.dragLayer
+          }}
+          />
+      </DraggableTarget>
+    );
   }
 }
 
@@ -185,22 +28,28 @@ Draggable.defaultPropTypes = {
 };
 
 Draggable.propTypes = {
-  /** callback when item was dropped in a new location */
-  onDrop: PropTypes.func,
-  /** callback when item is hovered*/
-  onHover: PropTypes.func,
   /** a function to render each item in the list */
   render: PropTypes.func.isRequired,
   /** decide whether to render a handle using `connectHandle` (see below) */
   withHandle: PropTypes.bool,
-  /** uniq id of list that contain current draggable item */
+  /** uniq id of container that contain current draggable item */
   containerId: PropTypes.string,
+  /* name of group between inside of each dnd is allowed */
   groupName: PropTypes.string,
+  /* custom renderer for item */
   renderItem: PropTypes.func,
+  /* position of item in container items array */
   index: PropTypes.number,
+  /* uniq id of an item */
   id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  /* model that represent item */
   item: PropTypes.object,
-  onMoveOut: PropTypes.func
+  /** callback when item was moved out from current container to another container */
+  onMoveOut: PropTypes.func,
+  /** callback when item was dropped in a new location */
+  onDrop: PropTypes.func,
+  /** callback when item is hovered*/
+  onHover: PropTypes.func
 };
 
 export default Draggable;
