+++ src/DragAndDrop/Draggable/types.js
@@ -1,0 +1,3 @@
+export const ItemTypes = {
+  DRAGGABLE: 'Draggable'
+};
